pyproject.toml
[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[project]
name = "fast_chat"
version = "0.1.0"
description = ""
authors = [{ name = "justyfay", email = "><sakura.ustinya@yandex.ru" }]
requires-python = ">=3.12,<4"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "fastapi>=0.115.2,<0.116",
    "python-jose[cryptography]>=3.3.0,<4",
    "passlib[argon2]>=1.7.4,<2",
    "python-multipart>=0.0.12,<0.0.13",
    "loguru>=0.7.2,<0.8",
    "pydantic-settings>=2.6.0,<3",
    "sqlalchemy>=2.0.36,<3",
    "alembic>=1.13.3,<2",
    "asyncpg>=0.30.0,<0.31",
    "greenlet>=3.1.1,<4",
    "pydantic[email]>=2.9.2,<3",
    "jinja2>=3.1.4,<4",
    "websockets>=13.1,<14",
    "uvicorn[standard]>=0.32.0,<0.33",
    "redis>=5.2.0,<6",
    "aiogram>=3.13.1,<4",
    "celery>=5.4.0,<6",
    "gunicorn>=23.0.0,<24",
    "flower>=2.0.1,<3",
    "ruff>=0.8.3,<0.9",
]
